pyproject.toml
[tool.poetry]
name = "final-project-server"
version = "0.1.0"
description = "boostcamp final project backend server"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
black = "^22.3.0"
pandas = "^1.4.2"
pydantic = "^1.9.1"
easydict = "^1.9"
PyMySQL = "^1.0.2"
PyYAML = "^6.0"
numpy = "^1.22.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
